backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
whitenoise==6.6.0